+++ source/dotnet/domain-tests/DomainTests.csproj
@@ -45,7 +45,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\wholesale-api\Application\Application.csproj" />
+    <ProjectReference Include="..\wholesale-api\Common\Common.csproj" />
   </ItemGroup>
 
 </Project>
